Move selection logic to asset grid class
